Documentation / examples / general refactor
Details
. Improved documentation (public api docs and README.md);
. Improved examples (minimal and elaborate).

TODO:
. Patch "send" method as "01-minimal" is using it and we can clearly see that it is not working as there is no "body" on markdown result;
. Think of a better way to test response middleware.
