+++ examples/01-minimal/index.js
@@ -1,0 +1,9 @@
+const { server } = require('./src/server');
+
+(async () => {
+  try {
+    await server.start();
+  } catch(err) {
+    console.error(err)
+  }
+})();
