+++ examples/01-minimal/src/__helpers__/close-api-opened-on-random-port.js
@@ -1,0 +1,3 @@
+const { server } = require('../server');
+
+exports.closeApiOpenedOnRandomPort = (t) => server.close(t.context.api);
