+++ examples/02-elaborate/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "elaborate",
+  "version": "1.0.0",
+  "description": "More elaborate functional test setup with API doc creation.",
+  "main": "index.js",
+  "scripts": {
+    "start": "PORT=3000 NODE_ENV=development nodemon index.js",
+    "test": "PORT=8080 NODE_ENV=test ava ---verbose --serial",
+    "test:create-docs": "CREATE_DOCS=true npm test",
+    "test:log-messages": "LOG_MESSAGES=true npm test",
+    "test:log-redux": "LOG_REDUX=true npm test"
+  },
+  "author": "Leonardo Sarmento de Castro",
+  "license": "ISC",
+  "devDependencies": {
+    "ava": "^1.4.1",
+    "body-parser": "^1.18.3",
+    "chalk": "^2.4.2",
+    "detect-port": "^1.3.0",
+    "express": "^4.16.4",
+    "got": "^9.6.0",
+    "nodemon": "^1.18.11",
+    "the-owl": "../../"
+  },
+  "dependencies": {}
+}
