+++ examples/02-elaborate/src/modules/__helpers__/close-api-opened-on-random-port.js
@@ -1,0 +1,3 @@
+const { server } = require('../../server');
+
+exports.closeApiOpenedOnRandomPort = (t) => server.close(t.context.api);
