+++ examples/02-elaborate/src/modules/__helpers__/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  ...require('./close-api-opened-on-random-port'),
+  ...require('./start-api-on-random-port'),
+};
