+++ examples/02-elaborate/src/modules/users/__fixtures__/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ...require('./users'),
+};
