+++ examples/02-elaborate/src/modules/users/__fixtures__/users.js
