+++ examples/02-elaborate/src/modules/users/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ...require('./router'),
+};
