+++ examples/02-elaborate/src/modules/users/router.js
@@ -1,0 +1,8 @@
+const { usersController } = require('./controller');
+
+exports.usersRouter = {
+  connect(app) {
+    app.get('/users/:id', usersController.findById);
+    app.post('/users', usersController.createUser);
+  },
+};
