+++ examples/02-elaborate/src/server/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  ...require('./connect'),
+  ...require('./messages'),
+  ...require('./server'),
+};
