--- examples/using-express-ava/README.md
@@ -1,27 +1,0 @@
-# using-express-ava
-
-Example project how to setup `the-owl` with Express.js and AVA test runner.
-
-
-## Where should I look?
-
-Inside `src/server/` folder:
-
-* `connect.js` shows how to connect the middleware with your Express.js app
-* `server.js` wraps Express.js in order to give it the ability to start/close the server on functional tests
-
-Inside `src/modules/user` folder:
-
-* `_fixtures.js` plain objects that will be served on routes (to "simulate a database")
-* `controller.js` has the implementation for `/users/:id` route
-* `router.js` map a route to a controller method
-* `__tests__` is where the gold is.
-
-
-## How can I see docs being generated?
-
-By running the script:
-
-```sh
-npm run test:create-docs
-```
