--- examples/using-express-ava/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "using-express-ava",
-  "version": "1.0.0",
-  "description": "Usage example with Express.js server and AVA test runner.",
-  "main": "index.js",
-  "scripts": {
-    "start": "PORT=8080 NODE_ENV=development nodemon index.js",
-    "test": "PORT=8081 NODE_ENV=test ava --serial",
-    "test:create-docs": "PORT=8081 NODE_ENV=test CREATE_DOCS=true ava",
-    "test:log-messages": "PORT=8081 NODE_ENV=test LOG_MESSAGES=true ava",
-    "test:log-redux": "PORT=8081 NODE_ENV=test LOG_REDUX=true ava"
-  },
-  "author": "Leonardo Sarmento de Castro",
-  "license": "ISC",
-  "devDependencies": {
-    "ava": "^1.4.1",
-    "axios": "^0.18.0",
-    "body-parser": "^1.18.3",
-    "chalk": "^2.4.2",
-    "cors": "^2.8.5",
-    "errorhandler": "^1.5.0",
-    "express": "^4.16.4",
-    "morgan": "^1.9.1",
-    "nodemon": "^1.18.11",
-    "the-owl": "../../"
-  },
-  "dependencies": {}
-}
