--- examples/using-express-ava/src/modules/users/index.js
@@ -1,2 +1,0 @@
-exports.usersController = require('./controller');
-exports.usersRouter = require('./router');
