--- examples/using-express-ava/src/modules/users/router.js
@@ -1,10 +1,0 @@
-const usersController = require('./controller');
-
-const usersRouter = {
-  connect(app) {
-    app.get('/users/:id', usersController.findById);
-    app.post('/users', usersController.createUser);
-  }
-};
-
-module.exports = usersRouter;
