--- examples/using-express-ava/src/server/index.js
@@ -1,3 +1,0 @@
-exports.connect = require('./connect');
-exports.messages = require('./messages');
-exports.server = require('./server');
