+++ package.json
@@ -37,7 +37,6 @@
   ],
   "devDependencies": {
     "ava": "^1.4.1",
-    "axios": "^0.19.0",
     "nyc": "^14.1.0"
   },
   "dependencies": {
